requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# YouTube Transcript
youtube-transcript-api==0.6.2

# AI/ML
openai==1.54.0

# Data Validation & Models
pydantic==2.9.2
pydantic-settings==2.6.0

# Environment & Config
python-dotenv==1.0.1

# HTTP Client (for additional features)
httpx==0.27.2

# Utilities
python-multipart==0.0.9  # For file uploads if needed later

# Development & Testing (optional but recommended)
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
isort==5.13.2